problem: jeromq is not thread safe
Solution: port CLIENT-SERVER from libzmq, which are thread safe sockets.

The thread safe sockets family include other sockets, which are not part of this commit.

Also, thread-safe sockets cannot be used with a poller at the moment. To use with a poller zmq_poller from libzmq should be ported as well.